HW3: rename models folder to moduls
